requirements-dev.txt
#
# This file is autogenerated by pip-compile
# Make changes in requirements-dev.in, then run this to update:
#
#    pip-compile requirements-dev.in
#
-e .
alabaster==0.7.7          # via sphinx
babel==2.2.0              # via sphinx
click==6.3                # via pip-tools
django-appconf==1.0.1
docutils==0.12            # via sphinx
first==2.0.1              # via pip-tools
flake8==2.5.4
isort==4.2.2
Jinja2==2.8               # via sphinx
MarkupSafe==0.23          # via jinja2
mccabe==0.4.0
pep257==0.7.0
pep8-naming==0.3.3
pep8==1.7.0               # via flake8
pip-tools==1.6
pluggy==0.3.1             # via tox
py==1.4.31                # via pytest, pytest-capturelog, tox
pyenchant==1.6.6
pyflakes==1.0.0           # via flake8
Pygments==2.1.3           # via sphinx
pytest-capturelog==0.7
pytest-django==2.9.1
pytest==2.9.0
pytz==2015.7              # via babel
six==1.10.0               # via django-appconf, pip-tools, sphinx, sphinxcontrib-spelling
snowballstemmer==1.2.1    # via sphinx
sphinx-rtd-theme==0.1.9   # via sphinx
sphinx==1.3.6
sphinxcontrib-spelling==2.1.2
tox==2.3.1
virtualenv==15.0.0        # via tox
